requirements.txt
openai>=0.10.0
sentence-transformers>=2.0.0
scikit-learn>=0.24
python-dotenv>=0.19.0
numpy>=1.19.2
flask>=2.0.0
flask-cors>=3.0.10
beautifulsoup4>=4.9.3
requests>=2.26.0

